requirements.txt
pandas>=1.5.0
numpy>=1.20.0
statsmodels>=0.13.0
scipy>=1.9.0
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
torch>=2.0.0
tqdm>=4.60.0
pydantic>=2.0.0
spacy>=3.7.0
nltk>=3.8.0
